adding common log attributes to be accessible for view boundary logs
